Remove erroneous casts of list item offsets
The offset of an item in a list is read and written as an int,
but for some reason we were casting it down to a short when
recording the offsets during list building.

Reviewed-on: https://asterix-gerrit.ics.uci.edu/292
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ asterix-om/src/main/java/edu/uci/ics/asterix/builders/AbstractListBuilder.java
@@ -77,7 +77,7 @@ public abstract class AbstractListBuilder implements IAsterixListBuilder {
     public void addItem(IValueReference item) throws HyracksDataException {
         try {
             if (!fixedSize && (item.getByteArray()[0] != serNullTypeTag || itemTypeTag == ATypeTag.ANY))
-                this.offsets.add((short) outputStorage.getLength());
+                this.offsets.add(outputStorage.getLength());
             if (itemTypeTag == ATypeTag.ANY
                     || (itemTypeTag == ATypeTag.NULL && item.getByteArray()[0] == serNullTypeTag)) {
                 this.numberOfItems++;
